// src/app/swaggerFiles/model/eleve.ts
/**
 * assoparenteleve API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Eleve { 
    id?: number;
    nom?: string;
    prenom?: string;
    dateNaissance?: string;
    niveau?: string;
}